Split visibility, light, audio logic into a separate file
Also standardized imports and removed unused imports.
